backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cors==1.0.1
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
future==0.18.3
gevent==23.9.1
greenlet==3.0.3
idna==3.6
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.3
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.4
tldextract==5.1.1
tzdata==2023.4
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.1.0
voluptuous==0.14.1
zope.event==5.0
zope.interface==6.1
